Add: QueueListener to store tweets in a queue
Once a stream to Twitter has been established, we have to read data from
the stream. The `QueueListener` class handles data read from the stream
by putting it into an internal queue. The class itself is iterable,
which makes taking data from the queue easy. This work is progress
towards gh-1.

The data is kept in (JSON) string format, and must be parsed into a
Tweet object separately. This decision was made so that the original
data string is available to archive later, such as to a file.