Add: Stream class to connect to Twitter
- Added `Stream` class. This closes gh-1.
- Can use a `ConfigParser` object for credentials. This is progress
towards gh-2.

The `Stream` class allows us to connect to the Twitter Streaming API by
connecting to either the `sample` or `filter` endpoint.

In order to work, you must provide authentication credentials provided
by Twitter. First, go to http://dev.twitter.com and create an app. The
consumer key and secret will be generated for you after, for use as
`consumer_key` and `consumer_secret`. After the previous step, you will
be redirected to your app's page. Create an access token under the the
"Your access token" section, for use as `access_token` and
`access_token_secret`.

The `Stream` class can be called with the credentials, or passed a
`ConfigParser` object with the appropriate credentials under a "twitter"
section. The configuration file itself needs to be parsed before
creating the `Stream`.