Task: have all classifiers subclass from an ABC
An abstract base class (ABC) won't allow a subclass to be instantiated
unless it implements any and all abstract methods or properties. This
allows code to fail more quickly, instead of silently until actually
calling an abstract method. In addition, by subclassing we are able to
potentially reuse common methods for classes of the same type.

With this new scheme, it makes sense to put our classifiers inside of a
single module. Therefore, we are also renaming the `naive_bayes.py` file
to `classify.py`.