Add: evaluate performance of a classifier
Given a gold standard reference and the labels of a test set, we can
compute various metrics of how well a classifier performs. The
performance metrics that are computed include:

- confusion matrix
- accuracy (overall)
- recall (per class, average, and weighted average)
- precision (per class, average, and weighted average)
- F-measure (per class, average, and weighted average)

While this doesn't use `matplotlib` directly, this provides the
foundation for analyzing how experimental variables affect the
performance of a classifier. This closes gh-10.