Task: move unit test docstring to comment
If a unit test has a docstring then it is used instead of the full call
of the unit test itself. Having the path to the unit test is useful for
figuring out exactly which test failed, so we are converting the
docstring to a regular comment in order to get that information back.
[ci skip]