Fix ZeroDivisionError during conversion from log
If the log-likelihood is too small, when we convert back using
`math.exp`, the result will round to zero. Luckily we can normalize the
log-probabilities by moving them all closer to zero by the same amount.
Doing so allows helps prevent rounding errors causing them all to round
to zero, and doesn't affect the conversion to a probability `[0--1]`.
This fixes gh-14.