Add parsing for new SemEval-2013 data format
The new development and test sets (both Twitter and SMS) are in a
different format than the publicly released training and development
files.

The new ones sometimes have use strings instead of integers for the
`sid` and `uid` fields. In addition, the `text` field appears to have
been written using Python's `repr()` function, causing quotes to be
escaped (`\"`) in the text, as well as other escape sequences
(especially Unicode) being present. Therefore, these need to be parsed
before use, and the Unicode is converted to an ASCII equivalent. When
parsing the test-dataset, the label "unknwn" is converted to `None`.
This closes gh-15.