API for MultinomialNB.classify return confidence
Previously the `classify` function only returned the most confident
class label. If we also wanted to know the confidence of the label, we
would have to make another call that would result in a recalculation of
the same probability. Therefore, we now return both the label and the
confidence for the label at the same time, and leave it up to the caller
to keep or discard what they wish.