Split project: library portion now named infer
During development, it became clear that we were working on what should
be two separate projects. One is a toolkit for machine learning and
experimentation. The other uses that toolkit for a specific task and in
a particular domain (Twitter).

The toolkit or library will be renamed to "infer", while the project
centered around Twitter data will retain the original name "infertweet".