Add dependency requirements for the infer library
After splitting the repository, we now need to put a dependency on the
library portion that was split off. Currently we do this by referring to
the specific git commit from GitHub, which allows us not to pollute PyPI
with many development releases. However, time will tell whether or not
we should use PyPI instead, even for daily development releases.