Add web user interface
We have added a web user interface so that the system can be queried.
This closes gh-8.

Because the sentiment classifier takes a very long time to load
(somewhere around five minutes), we don't want to have to reload the
classifier each time we make a change to the web interface (the web
server needs to restart in order to apply any changes). Therefore, we
use a RPC server to load the classifier once and serve it as needed.
Note: we tried using an XML-RPC server first, but because our classifier
returns non-primitive data types (in this case a `namedtuple`), the XML
server could not "marshal" the return object. Therefore we used a Python
only implementation that I believe utilizes Python's pickle module to
send and receive objects.

In addition, we have added a configuration file that can be used to
change the ports and filenames the modules use.