Task: conditional probability for single features
For some reason, when classifying a document that consists only of one
feature, the hierarchical classifier only labels the document as neutral
or positive, even when the confidence values are less than 0.5.

I'm still not sure why this is the case, however I am addressing the
issue for the part of the web interface that shows the influence of each
of the individual features that make up the query by just using the
conditional probability of the feature across the labels instead of
trying to classify it. In the mean time, this addresses issue gh-27.