Add: use keywords to search Twitter for tweets
Instead of typing in the text of a single tweet to be classified, users
may enter keywords, which will be used to search for matching tweets
from Twitter that will then be classified and displayed to the user.
This is progress towards gh-9.

Currently we use a heuristic to determine if the query string is a
document to be classified or a set of keywords to be used for searching.
To be considered keywords, the query must be 3 words or less and must be
less than 30 characters, otherwise we assume it is a single document.