Remove logging of label, probability from query
Since we are expecting a most of the requests that come in to be keyword
queries, it doesn't make much sense to log the label and probability of
the keywords themselves. However, we are still interested in logging the
query itself, as that usage information is valuable.

In addition, we were also getting an `UnboundLocalError` exception most
likely due to a scoping issue. Since we no longer use the variables that
were causing the error, this also fixes gh-39.