Add: RESTful JSON API for sentiment analysis
Here we add the ability to programmatically request to classify a single
document (tweet) for sentiment analysis. This is progress towards gh-25.

Resource URL:
http://.../api/sentiment/service_name.json

Currently, we only accept GET requests, and 'classify' is the only valid
service name. In the future, we can add other services under different
names, such as a bulk classification service.