Move sentiment experiment code to another module
When we load the pickled classifier, we also need to load the original
tokenizer function that was used with it. Previously, when importing the
tokenizer we also indirectly caused other imports to be called, such as
for matplotlib. We don't want to require matplotlib to be loaded in
order to load the pickled classifier (especially if we're trying to use
PyPy). Therefore, we moved some of the experimentation code to another
module so that when we import the tokenizer function it doesn't cause
matplotlib to be imported as well.