Fix: erroneous neutral conditional probability
In an attempt to visually convey feature conditional probability to the
user, we color code instances according to their class label and the
confidence in that label. When calculating which features were to be
shown as gray (neutral), and which to show with color (subjective), we
erroneously returned the opposite label (and an incorrect value for
those we had been returning as neutral). This commit fixes those errors.