Use logging module instead of log process
We remove our custom process that we used to log queries, and instead
switch to the logging module, which Tornado uses by default. This allows
us to easily log the same messages to console and to a file. In
addition, we use the current system's hostname as part of the filename
for the log file.

However, doing this causes logging to a file to occur in the main
Tornado process, which has the potential to block the IO-loop.