Add: user feedback on output of the classifier
This adds the ability for a user to report a tweet as being
misclassified by allowing the user to flag the tweet with the correct
class label. Currently this logs the report to the log file, although
this could be expanded in the future to allow immediate training from
the newly labeled instance.

While this provides most of the work for gh-38, we are currently
missing:

- Allow reports for single document classifications (rather than
allowing only when the results are tweets).
- Have the server verify the report is not obviously invalid by
classifying the text and making sure the flag differs. The currently
assigned class label would also be beneficial to store in the log.
- If in the future we implement self-training, we need to *unlearn* any
flagged reports.