Add: estimate country from latitude and longitude
When our only data is latitude and longitude points and we want to
instead have a list of countries, we need some way to convert. The
method employed here estimates the country based on the shortest
Euclidean distance between the point and the average latitude and
longitude of each country.

There are online APIs that provide this as a service, such as
geonames.org, however using an offline method results in much faster
response times.