Change result-type from Twitter to most-recent
Previously the results that were returned by Twitter was a mix of the
most recent and also the most popular results. Unfortunately, we tend to
get a lot of repeated results when a popular tweet has many ReTweets in
a short amount of time. Therefore, we have switched our result-type to
only request the most recent results.