Add: active learning using uncertainty sampling
Allow users to help label data selected by an active learning algorithm,
in this case using uncertainty sampling. This closes gh-36.

Currently we retrieve 100 tweets from Twitter and sort them by the
confidence in their predicted labels. We then return the top-10 most
uncertain tweets, and allow the user to report a label.

The user is currently able to label as few or as many tweets as they
wish from the list returned by the system. It is worth thinking about if
this is the right approach, or if the system should generate a new list
after a certain amount of labels have been given (even a single report).