Use minimum objective or subjective probability
When selecting instances for active learning, we may be very sure that
the instance is positive or negative, but not so sure if the instance is
objective or subjective. This shouldn't usually be the case, but in case
there is this discrepancy we should have those instances rated uncertain
as well.