Add: location sentiment stats using Google Charts
When the geo-option is specified, the average sentiment for each country
is calculated and displayed on a map using the geochart visualization
from Google Charts. This closes gh-44.

Improvements can still be made. First, the number of geocoded tweets
returned for each keyword search is very low, therefore even popular
terms have little data. Additional pages of results may be retrieved,
but this will use many API calls and may result in being rate limited.
The second current drawback is the estimation of country from the geo-
coordinates.